refactor(plugin): no more mixins
